+++ src/main/java/com/cburch/logisim/std/memory/ShiftRegisterLogger.java
@@ -14,9 +14,9 @@ public class ShiftRegisterLogger extends InstanceLogger {
     @Override
     public Object[] getLogOptions(InstanceState state) {
         Integer stages = state.getAttributeValue(ShiftRegister.ATTR_LENGTH);
-        Object[] ret = new Object[stages.intValue()];
+        Object[] ret = new Object[stages];
         for (int i = 0; i < ret.length; i++) {
-            ret[i] = Integer.valueOf(i);
+            ret[i] = i;
         }
         return ret;
     }
